Move HostProcessReloadHandler to the correct package
