+++ host-controller/src/main/java/org/jboss/as/domain/controller/operations/HostProcessReloadHandler.java
@@ -19,7 +19,7 @@
 * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
 * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
 */
-package org.jboss.as.domain.controller.operations.deployment;
+package org.jboss.as.domain.controller.operations;
 
 import java.util.Locale;
